+++ Raterization/Raterization/inc/LMath.h
@@ -4,20 +4,20 @@
 #include <math.h>
 #include "Vector.h"
 
-const double PI = ((double)3.1415926535f);
-const double PI2 = ((double)6.283185307f);
-const double PI_DIV_2 = ((double)1.570796327f);
-const double PI_DIV_4 = ((double)0.785398163f);
-const double PI_INV = ((double)0.318309886f);
+const float PI = ((float)3.1415926535f);
+const float PI2 = ((float)6.283185307f);
+const float PI_DIV_2 = ((float)1.570796327f);
+const float PI_DIV_4 = ((float)0.785398163f);
+const float PI_INV = ((float)0.318309886f);
 
-inline double Angle_to_Radian(double angle)
+inline float Angle_to_Radian(float angle)
 {
-	return (angle * PI / 180.0);
+	return (float)(angle * PI / 180.0);
 }
 
-inline double Radian_to_Angle(double radian)
+inline float Radian_to_Angle(float radian)
 {
-	return (radian * 180.0 / PI);
+	return (float)(radian * 180.0 / PI);
 }
 
 enum 
@@ -28,12 +28,12 @@ enum
 	LINE_INTERSECT_EVERYWHERE
 };
 
-extern double sin_table[361];
-extern double cos_table[361];
+extern float sin_table[361];
+extern float cos_table[361];
 
 void Build_SinCos_Tables();
-double Fast_sin(double theta);
-double Fast_cos(double theta);
+float Fast_sin(float theta);
+float Fast_cos(float theta);
 
 struct Line2D
 {
@@ -48,8 +48,8 @@ struct Line2D
 
 	// basic methods
 	void init(Point2D pb, Point2D pe, Vector2D v);
-	Point2D compute(double t);
-	int intersect(Line2D* line, double *t1, double *t2);
+	Point2D compute(float t);
+	int intersect(Line2D* line, float *t1, float *t2);
 	int intersect(Line2D* line, Point2D* point);
 };
 typedef Line2D* PLine2D;
@@ -67,7 +67,7 @@ struct Line3D
 
 	// basic methods
 	void init(Point3D pb, Point3D pe, Vector3D v);
-	Point3D compute(double t);
+	Point3D compute(float t);
 };
 typedef Line3D* PLine3D;
 
@@ -88,8 +88,8 @@ struct Plane3D
 	// basic methods
 	void zero();
 	void init(Point3D point, Vector3D n, int normalize);
-	double compute(Point3D* point);
-	int intersect(Line3D* line, double* t, Point3D* point);
+	float compute(Point3D* point);
+	int intersect(Line3D* line, float* t, Point3D* point);
 
 	// override operator
 	Plane3D& operator=(const Plane3D& p);
@@ -97,19 +97,19 @@ struct Plane3D
 
 struct Polar2D
 {
-	double r, theta;
+	float r, theta;
 	
 	// Constructor
 	Polar2D() = default;
 	~Polar2D() = default;
 	Polar2D(const Polar2D& p) { *this = p; }
-	Polar2D(double dr, double dt) : r(dr), theta(dt) {}
+	Polar2D(float dr, float dt) : r(dr), theta(dt) {}
 	
 	// basic methods
-	void init(double dr, double dt);
+	void init(float dr, float dt);
 	void init(Point2D* point);
 	Point2D toPoint2D();
-	void toXY(double *dx, double *dy);
+	void toXY(float *dx, float *dy);
 
 	// override operator
 	Polar2D& operator=(const Polar2D& p);
@@ -118,19 +118,19 @@ typedef Polar2D* PPolar2D;
 
 struct Cylindrical3D
 {
-	double r, theta, z;
+	float r, theta, z;
 
 	// Constructor
 	Cylindrical3D() = default;
 	~Cylindrical3D() = default;
 	Cylindrical3D(const Cylindrical3D& c) { *this = c; }
-	Cylindrical3D(double dr, double dt, double dz) : r(dr), theta(dt), z(dz) {}
+	Cylindrical3D(float dr, float dt, float dz) : r(dr), theta(dt), z(dz) {}
 
 	// basic methods
-	void init(double dr, double dt, double dz);
+	void init(float dr, float dt, float dz);
 	void init(Point3D* point);
 	Point3D toPoint3D();
-	void toXYZ(double *dx, double *dy, double *dz);
+	void toXYZ(float *dx, float *dy, float *dz);
 
 	// override operator
 	Cylindrical3D& operator=(const Cylindrical3D& c);
@@ -139,19 +139,19 @@ typedef Cylindrical3D* PCylindrical3D;
 
 struct Spherical3D
 {
-	double pos, theta, phi;
+	float pos, theta, phi;
 
 	// Contructor
 	Spherical3D() = default;
 	~Spherical3D() = default;
 	Spherical3D(const Spherical3D& s) { *this = s; }
-	Spherical3D(double dpos, double dt, double dphi) : pos(dpos), theta(dt), phi(dphi) {}
+	Spherical3D(float dpos, float dt, float dphi) : pos(dpos), theta(dt), phi(dphi) {}
 
 	// basic methods
-	void init(double dpos, double dt, double dphi);
+	void init(float dpos, float dt, float dphi);
 	void init(Point3D* point);
 	Point3D toPoint3D();
-	void toXYZ(double *dx, double *dy, double *dz);
+	void toXYZ(float *dx, float *dy, float *dz);
 
 	// override operator
 	Spherical3D& operator=(const Spherical3D& s);
@@ -164,7 +164,7 @@ inline void Line2D::init(Point2D pb, Point2D pe, Vector2D v)
 	p0 = pb; p1 = pe; vec = v;
 }
 
-inline Point2D Line2D::compute(double t)
+inline Point2D Line2D::compute(float t)
 {
 	Point2D tmp;
 	tmp.x = p0.x + vec.x * t;
@@ -172,7 +172,7 @@ inline Point2D Line2D::compute(double t)
 	return tmp;
 }
 
-inline int Line2D::intersect(Line2D* line, double *t1, double *t2)
+inline int Line2D::intersect(Line2D* line, float *t1, float *t2)
 {
 	auto det_line = (vec.x * line->vec.y - vec.y * line->vec.x);
 	if (abs(det_line) <= EPSILON_E5)
@@ -212,7 +212,7 @@ inline void Line3D::init(Point3D pb, Point3D pe, Vector3D v)
 	p0 = pb; p1 = pe; vec = v;
 }
 
-inline Point3D Line3D::compute(double t)
+inline Point3D Line3D::compute(float t)
 {
 	Point3D tmp;
 	tmp.x = p0.x + vec.x * t;
@@ -233,14 +233,14 @@ inline void Plane3D::init(Point3D point, Vector3D n, int normalize)
 	if (normalize) normal.normalize();
 }
 
-inline double Plane3D::compute(Point3D* point)
+inline float Plane3D::compute(Point3D* point)
 {
 	return (normal.x * (point->x - pos.x) +
 		normal.y * (point->y - pos.y) +
 		normal.z * (point->z - pos.z));
 }
 
-inline int Plane3D::intersect(Line3D* line, double* t, Point3D* point)
+inline int Plane3D::intersect(Line3D* line, float* t, Point3D* point)
 {
 	auto dot = normal.dot(&line->vec);
 	if (abs(dot) <= EPSILON_E5)
@@ -270,7 +270,7 @@ inline Plane3D& Plane3D::operator=(const Plane3D& p)
 }
 
 // Polar2D methods implement
-inline void Polar2D::init(double dr, double dt)
+inline void Polar2D::init(float dr, float dt)
 {
 	r = dr; theta = dt;
 }
@@ -289,7 +289,7 @@ inline Point2D Polar2D::toPoint2D()
 	return tmp;
 }
 
-inline void Polar2D::toXY(double *dx, double *dy)
+inline void Polar2D::toXY(float *dx, float *dy)
 {
 	*dx = r * cos(theta);
 	*dy = r * sin(theta);
@@ -302,7 +302,7 @@ inline Polar2D& Polar2D::operator=(const Polar2D& p)
 }
 
 // Cylindrical3D methods implement
-inline void Cylindrical3D::init(double dr, double dt, double dz)
+inline void Cylindrical3D::init(float dr, float dt, float dz)
 {
 	r = dr; theta = dt; z = dz;
 }
@@ -323,7 +323,7 @@ inline Point3D Cylindrical3D::toPoint3D()
 	return tmp;
 }
 
-inline void Cylindrical3D::toXYZ(double *dx, double *dy, double *dz)
+inline void Cylindrical3D::toXYZ(float *dx, float *dy, float *dz)
 {
 	*dy = r * cos(theta);
 	*dz = r * sin(theta);
@@ -337,7 +337,7 @@ inline Cylindrical3D& Cylindrical3D::operator=(const Cylindrical3D& c)
 }
 
 // Spherical3D methods implement
-inline void Spherical3D::init(double dpos, double dt, double dphi)
+inline void Spherical3D::init(float dpos, float dt, float dphi)
 {
 	pos = dpos; theta = dt; phi = dphi;
 }
@@ -352,7 +352,7 @@ inline void Spherical3D::init(Point3D* point)
 inline Point3D Spherical3D::toPoint3D()
 {
 	Point3D tmp;
-	double r;
+	float r;
 	r = pos * sin(phi);
 	tmp.z = pos * cos(phi);
 
@@ -362,9 +362,9 @@ inline Point3D Spherical3D::toPoint3D()
 	return tmp;
 }
 
-inline void Spherical3D::toXYZ(double *dx, double *dy, double *dz)
+inline void Spherical3D::toXYZ(float *dx, float *dy, float *dz)
 {
-	double r;
+	float r;
 	r = pos * sin(phi);
 	*dz = pos * cos(phi);
 
@@ -380,20 +380,20 @@ inline Spherical3D& Spherical3D::operator=(const Spherical3D& s)
 
 // Transform functions
 // 2 Transform
-inline void Point2D_To_RTheta(PPoint2D point, double *r, double *theta)
+inline void Point2D_To_RTheta(PPoint2D point, float *r, float *theta)
 {
 	*r = sqrt((point->x * point->x) + (point->y * point->y));
 	*theta = atan(point->y / point->x);
 }
 
-inline void Point3D_To_RThetaZ(PPoint3D point, double *r, double *theta, double *z)
+inline void Point3D_To_RThetaZ(PPoint3D point, float *r, float *theta, float *z)
 {
 	*r = sqrt((point->x * point->x) + (point->y * point->y));
 	*theta = atan(point->y / point->x);
 	*z = point->z;
 }
 
-inline void Point3D_To_PosThetaPhi(PPoint3D point, double *pos, double *theta, double *phi)
+inline void Point3D_To_PosThetaPhi(PPoint3D point, float *pos, float *theta, float *phi)
 {
 	*pos = sqrt((point->x * point->x) + (point->y * point->y) + (point->z * point->z));
 	*theta = atan(point->y / point->x);
